refacto: prop drilling
