+++ src/components/UserProfile/ModalStyling/ModalStyling.js
@@ -25,24 +25,29 @@ const buttonVariants = {
 
 const Button = styled.button(
   {
-    padding: "10px 15px",
+    marginTop: "20px",
+    padding: "10px",
     border: "0",
     lineHeight: "1",
     borderRadius: "3px",
+    width: "100%",
   },
   ({ variant = "primary" }) => buttonVariants[variant]
 );
 
 const Input = styled.input({
-  borderRadius: "3px",
-  border: `1px solid ${colors.gray10}`,
+  border: `1px solid ${colors.gray20}`,
+  padding: "10px",
   background: "#f1f2f7",
-  padding: "8px 12px",
+  borderRadius: "10px",
+  cursor: "pointer",
+  width: "100%",
 });
 
 const FormGroup = styled.div({
   display: "flex",
   flexDirection: "column",
+  justifyContent: "flex-end",
 });
 
 const CircleButton = styled.button({
@@ -61,13 +66,14 @@ const CircleButton = styled.button({
 });
 
 const Dialog = styled(ReachDialog)({
+  display: "block",
   maxWidth: "450px",
-  borderRadius: "3px",
+  borderRadius: "10px",
   paddingBottom: "3.5em",
   boxShadow: "0 10px 30px -5px rgba(0, 0, 0, 0.2)",
   margin: "20vh auto",
   [mq.small]: {
-    width: "100%",
+    width: "70%",
     margin: "10vh auto",
   },
 });
